oauth2_clientmanager: don't replace non-socket file with socket
If the specified pathname for the socket already exists, it will be
replaced regardless of whether it was already a socket.  This is probably
not expected behavior, so we'll bail with OSError instead if it exists
and isn't a socket.

+++ oauth2_clientmanager/__init__.py
@@ -17,6 +17,7 @@ import time
 import getpass
 import shutil
 import signal
+import stat
 
 from typing import cast, Any, Dict, Optional, Sequence, Tuple, Type, Union
 
@@ -564,7 +565,11 @@ class OAuth2ClientManager:
         self._debug(f"Starting HTTP listener on {filename}")
 
         if os.path.exists(filename):
-            os.unlink(filename)
+            sock_stat = os.stat(filename)
+            if stat.S_ISSOCK(sock_stat.st_mode):
+                os.unlink(filename)
+            else:
+                raise OSError("{filename} already exists but is not a socket.  Will not replace.")
         self._server = _UnixSocketThreadingHTTPServer(filename, _TokenSocketHandler, self)
         self._server.context = self
         os.chmod(filename, 0o600)
